+++ frontend/src/Components/Pages/GamePage.js
@@ -3,6 +3,7 @@ import StartScene from '../Game/Scenes/StartScene';
 import GameScene from '../Game/Scenes/GameScene';
 import GameOverScene from '../Game/Scenes/GameOverScene';
 import PauseGameScene from '../Game/Scenes/PauseGameScene';
+import PreloadScene from '../Game/Scenes/PreloadScene'
 
 let game;
 
@@ -17,7 +18,7 @@ const GamePage = () => {
       Audio:true
     },
   
-    scene: [StartScene, GameScene, GameOverScene,PauseGameScene],
+    scene: [PreloadScene,StartScene, GameScene, GameOverScene,PauseGameScene],
     pixelArt: true,
     physics: {
       default: "arcade",
